+++ src/web-pages/components.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+
+export default function Components() {
+    return (
+        <div className='nav-components'>
+            <ComponentsNav />
+        </div>
+        <div className='showdown-components'>
+            <ComponentsShowdown />
+        </div>
+    )
+}
